requirements.txt
# Core dependencies
tensorflow>=2.8.0
keras>=2.8.0
numpy>=1.21.0
h5py>=3.6.0
matplotlib>=3.5.0
scikit-image>=0.19.0
python-dotenv>=0.19.0

# Optional dependencies for video processing
opencv-python>=4.5.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
jupyter>=1.0.0